Double precision test
Testing out a double precision mode.

No clue how to integrate this into the suite.

+++ inc/gpu_roof.h
@@ -1,3 +1,3 @@
 #include "bench.h"
 
-void gpu_axpy(int, float, float, float *, float *, struct roof_args *);
+void gpu_axpy(int, double, double, double *, double *, struct roof_args *);
