+++ src/generic/gpu_roof.c
@@ -1,6 +1,6 @@
 #include "roof.h"
 
-void gpu_axpy(int n, float a, float b, float * x_in, float * y_in,
+void gpu_axpy(int n, double a, double b, double * x_in, double * y_in,
               struct roof_args *args)
 {
     args->runtime = 0.;
